updated dec to float

+++ Assets/Scripts/ACP_PayloadQD.cs
@@ -4,6 +4,7 @@ using UnityEngine;
 using TMPro;
 using UnityEngine.UI;
 using System;
+using UnityEngine.Rendering;
 
 public class ACP_PayloadQD : MonoBehaviour
 {
@@ -35,7 +36,7 @@ public class ACP_PayloadQD : MonoBehaviour
 
     public bool Added;
     public bool Initial;
-    public decimal OldMoment;
+    public float OldMoment;
 
     public GameObject OriginalPosition;
     public GameObject CurrentPosition;
@@ -52,10 +53,10 @@ public class ACP_PayloadQD : MonoBehaviour
     public GameObject OBJ;
     public GameObject OBJ2;
     public GameObject StartPos;
-    public decimal CombinedWt;
+    public float CombinedWt;
     public Vector3 FSOposition;
     public GameObject WeightWarning;
-    public decimal FSo;
+    public float FSo;
     public Vector3 CBadjustedPosition;
     public Vector3 DistanceVector;
     public float Distance;
@@ -160,8 +161,8 @@ public class ACP_PayloadQD : MonoBehaviour
             Distance *= constant;
             float moment = PalletWeight * Distance;
             Moment = (float)(Math.Round(moment, 0));
-            Payload.Moment += (decimal)Moment;
-            OldMoment = (decimal)Moment;
+            Payload.Moment += Moment;
+            OldMoment = Moment;
             Added = true;
             LocalDistance = (float)(Math.Round(Distance, 0)); // this is the FS of the CB
 
@@ -179,9 +180,9 @@ public class ACP_PayloadQD : MonoBehaviour
             Distance *= constant;
             Debug.Log("distance after constant adjustment is .." + Distance);
             float moment = PalletWeight * Distance;
-            Payload.Moment += (decimal)moment;
+            Payload.Moment += moment;
             Moment = (float)(Math.Round(moment, 0));
-            OldMoment = (decimal)Moment;
+            OldMoment = Moment;
             Added = true;
             LocalDistance = (float)(Math.Round(Distance, 0));
             GameObject.FindGameObjectWithTag("SelectedPanel").gameObject.transform.GetChild(3).gameObject.transform
