+++ Assets/Scripts/LoadPaxCL.cs
@@ -8,21 +8,21 @@ public class LoadPaxCL : MonoBehaviour
     public GameObject paxprefaboriginal;
     public bool bool1;
     public static Material colour;
-    public static decimal PaxWeight;
+    public static float PaxWeight;
     public Material red;
     public Material green;
     public Material blue;
     public Material cyan;
     public static GameObject FlightStation0;
-    public static decimal Distance;
+    public static float Distance;
     public GameObject paxprefabA;
     public static Material Colour;
     public static bool PaxWeight200;
     public static bool PaxWeight210;
     public static bool PaxWeight230;
     public static bool PaxWeight250;
-    public decimal ThisPaxWeight;
-    public decimal constant;
+    public float ThisPaxWeight;
+    public float constant;
 
 
 
@@ -33,7 +33,7 @@ public class LoadPaxCL : MonoBehaviour
         gameObject.tag = "selectable";
         colour = green;
         FlightStation0 = FlightStationZero.FS0.gameObject;
-        constant = 39.37006151790835M;
+        constant = 39.37006151790835f;
 
 
     }
@@ -107,10 +107,10 @@ public class LoadPaxCL : MonoBehaviour
             Payload.TotalPaxWt = Payload.TotalPaxWt + PaxWeight;
             Payload.TotalPayloadWt += PaxWeight;
 
-            Distance = (decimal)(transform.position.x - FlightStation0.transform.position.x);
+            Distance = (float)(transform.position.x - FlightStation0.transform.position.x);
             Distance = Distance * constant;
-            decimal moment = PaxWeight * Distance;
-            decimal paxint = (Math.Round(moment, 0));
+            float moment = PaxWeight * Distance;
+            float paxint = (float)Math.Round(moment, 0);
             Payload.Moment += paxint;
             mypaxprefab.GetComponent<DestroyPax>().ThisPassengerMoment = paxint;
 
@@ -133,11 +133,11 @@ public class LoadPaxCL : MonoBehaviour
             Payload.TotalPaxWt = Payload.TotalPaxWt + PaxWeight;
             Payload.TotalPayloadWt += PaxWeight;
 
-            Distance = (decimal)(transform.position.x - FlightStation0.transform.position.x);
+            Distance = (float)(transform.position.x - FlightStation0.transform.position.x);
             Distance = Distance * constant;
-            decimal BetterDistance = Math.Round(Distance, 0);
-            decimal moment = PaxWeight * BetterDistance;
-            decimal paxint = (Math.Round(moment, 0));
+            float BetterDistance = (float)Math.Round(Distance, 0);
+            float moment = PaxWeight * BetterDistance;
+            float paxint = (float)Math.Round(moment, 0);
             Payload.Moment += paxint;
             mypaxprefab.GetComponent<DestroyPax>().ThisPassengerMoment = paxint;
         }
