+++ Assets/Scripts/TPU.cs
@@ -10,10 +10,10 @@ public class TPU : MonoBehaviour
     
     public GameObject TPUprefab;
     public GameObject MainCanvas;
-    public  decimal weightint;
+    public  float weightint;
     public Button Btn;
     public TMP_InputField input;
-    public decimal   Dec;
+    public float   Dec;
     public int Int;
     public bool check;
 
@@ -32,7 +32,7 @@ public class TPU : MonoBehaviour
             GameObject TPUprefab1 = Instantiate(TPUprefab);
             TPUprefab1.transform.localPosition = new Vector3(0, 0.5f, 0);
             TPUprefab1.transform.localEulerAngles = new Vector3(-90, 0, 0);
-            weightint = Convert.ToDecimal(input.text);
+            weightint = (float)Convert.ToDecimal(input.text);
             PartOne.PartOneTotalWt += weightint; // this adds the weight to the payload panel
             Debug.Log("TPU ADDED");
